MEMOPT: cut live-stacking memory (~1 GB RSS on IMX533 sessions)
A colour live-stack session on a 9 MP OSC camera held the process at
~1 GB. Two compounding causes: ~310 MB of session residents (some of it
avoidable) and ~260 MB of large-object-heap arrays allocated and
discarded on EVERY integrated frame, which kept the heap inflated and
fragmented even under Workstation GC + GCConserveMemory.

Residents:
- The mono accumulator (float[N], 35 MB) was allocated in colour
  sessions but never written — now allocated only on the mono branch.
- The normalised master flat is stored as float[] instead of double[]
  (35 MB instead of 69 MB per master key; a flat lives in [~0.5, 2.0]
  where float error ~1e-7 is far below photon noise). Applies to both
  the live pre-processor cache and the STUDIO batch calibration cache.
- Reset() now runs a one-off LOH-compacting collection so a finished
  session actually returns its ~300 MB to the OS.

Per-frame churn -> session scratch (frame geometry is constant within a
session; only buffers that never escape AddFrameAsync are reused, all
nulled on Reset):
- BayerDebayer.Bilinear + ImageResampler.ApplyTransform +
  CalibrationMath.CalibratePixels gained destination-buffer overloads
  (legacy overloads delegate; batch/studio call sites unchanged). The
  resampler explicitly zeroes off-canvas pixels so a dirty reused
  buffer can't leak the previous frame.
- The cumulative-SNR reconstruction reuses scratch too — and colour
  sessions now reconstruct the Rec.601 luminance of the running mean
  instead of reading the never-written mono buffer, so colour cumulative
  SNR reports a real value instead of 0.
- StarDetector's visited bool[N] (9 MB/frame) and every per-partition
  int[65536] histogram (256 KB LOH each, x partitions x ~5 statistics
  passes per frame — tens of MB on a many-core host) are rented from
  ArrayPool instead of allocated.

Deployment parity: deploy/nina-polaris.service and the Dockerfile now
carry the same DOTNET_GCConserveMemory=5 / MALLOC_ARENA_MAX=2 /
MALLOC_TRIM_THRESHOLD_ fixes as the .deb unit (they were missing, so
those installs never got the glibc-arena RSS fix).

Tests: new MemoryOptimizationTests pin (a) bit-identical output of every
destination-buffer overload against the legacy allocating one, including
worst-case dirty destinations, (b) a colour end-to-end session (drifted
frames, warp path, no mono buffer, real cumulative SNR), and (c) an
allocation regression guard — per-frame managed allocation on a 1024^2
colour session dropped from 44.3 MB to 11.8 MB measured. Full suite:
only the 12 known pre-existing failures.

+++ src/NINA.Polaris/Services/Studio/CalibrationService.cs
@@ -105,8 +105,8 @@ public class CalibrationService {
                 return img;
             }
 
-            var flatCache = new Dictionary<(string flat, string? cal), (double[] norm, double mean)>();
-            (double[] norm, double mean) GetNormalizedFlat(string flatPath, string? calibratorPath) {
+            var flatCache = new Dictionary<(string flat, string? cal), (float[] norm, double mean)>();
+            (float[] norm, double mean) GetNormalizedFlat(string flatPath, string? calibratorPath) {
                 var key = (flatPath, calibratorPath);
                 if (flatCache.TryGetValue(key, out var cached)) return cached;
                 var flat = LoadMaster(flatPath);
@@ -170,7 +170,7 @@ public class CalibrationService {
     private void CalibrateOne(
             string lightPath, CalibrationRequest req, MasterIndex masters,
             Func<string, BaseImageData> loadMaster,
-            Func<string, string?, (double[] norm, double mean)> getNormFlat,
+            Func<string, string?, (float[] norm, double mean)> getNormFlat,
             string outRoot, string rigName) {
         if (!File.Exists(lightPath))
             throw new InvalidOperationException($"Light missing on disk: {lightPath}");
@@ -225,7 +225,7 @@ public class CalibrationService {
         // Dimension validation also lives there (throws on mismatch).
         ushort[]? darkPx = darkPath != null ? loadMaster(darkPath).Data : null;
         ushort[]? biasPx = (darkPath == null && biasPath != null) ? loadMaster(biasPath).Data : null;
-        (double[] norm, double mean)? flat = flatPath != null
+        (float[] norm, double mean)? flat = flatPath != null
             ? getNormFlat(flatPath, flatCalibrator)
             : null;
         var pixels = CalibrationMath.CalibratePixels(light.Data, darkPx, biasPx, flat);
